+++ sys/kern/vfs_vnops.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)vfs_vnops.c	7.44 (Berkeley) 7/6/92
+ *	@(#)vfs_vnops.c	7.45 (Berkeley) 7/7/92
  */
 
 #include "param.h"
@@ -342,7 +342,7 @@ vn_ioctl(fp, com, data, p)
 		if (com == FIONREAD) {
 			if (error = VOP_GETATTR(vp, &vattr, p->p_ucred, p))
 				return (error);
-			*(off_t *)data = vattr.va_size - fp->f_offset;
+			*(int *)data = vattr.va_size - fp->f_offset;
 			return (0);
 		}
 		if (com == FIONBIO || com == FIOASYNC)	/* XXX */
